Update README.md
When the user is registering himself to the webpage, he's writing e-mail, username, password and one input field which is for his balance. This balance is for example his own balance, which he has in his credit card and then when he's logged can deposit from the sum he has entered in the register page. Everything of that is saved in database. Also there are  sections for previous matches, next matches and matches today.  There is section "My bets" which is showing if the user is in his profile. Also the user can bet matches only if he's logged.